Geometry Test Refactor Part 2
Further split test_geometry
moved global variables into individual tests to ensure tests run in isolation
